+++ src/main/java/nl/delftelectronics/spaceinvaders/core/entities/MediumEnemy.java
@@ -20,7 +20,7 @@ public class MediumEnemy extends Enemy {
      * @param southBoundary   southernmost boundary of the playing field.
      * @param block 		  the block of enemies this enemy belongs to.
      */
-    public MediumEnemy(Integer positionX, Integer positionY, Integer width, Integer height,
+    public MediumEnemy(double positionX, double positionY, double width, double height,
                        Integer westBoundary, Integer eastBoundary, Integer southBoundary,
                        EnemyBlock block) {
         super(positionX, positionY, width, height, westBoundary, eastBoundary, southBoundary,
